fix(toast): toasts now render when prefersReducedMotion is enabled
